separate group node from node
